修改bug

+++ src/test/java/org/liahnu/bot/util/elo/CalculateEloTest.java
@@ -19,32 +19,32 @@ public class CalculateEloTest {
     public void test4BaseRule() {
 
         EloCalculateContext context = new EloCalculateContext();
-        Map<Long,BigDecimal> originalElo = Map.of(
-                1L, new BigDecimal("1000"),
-                2L, new BigDecimal("1000"),
-                3L, new BigDecimal("1000"),
-                4L, new BigDecimal("1000")
+        Map<Integer, BigDecimal> originalElo = Map.of(
+                1, new BigDecimal("1000"),
+                2, new BigDecimal("1000"),
+                3, new BigDecimal("1000"),
+                4, new BigDecimal("1000")
         );
-        Map<Long,BigDecimal> scores = Map.of(
-                1L, new BigDecimal("80"),
-                2L, new BigDecimal("-3.8"),
-                3L, new BigDecimal("-26.4"),
-                4L, new BigDecimal("-49.7")
+        Map<Integer, BigDecimal> scores = Map.of(
+                1, new BigDecimal("80"),
+                2, new BigDecimal("-3.8"),
+                3, new BigDecimal("-26.4"),
+                4, new BigDecimal("-49.7")
         );
 
         context.setOriginalElo(originalElo);
         context.setScores(scores);
 
-        Map<Long, BigDecimal> calculate = EloCalculate.calculate(ContestType.RCR, context);
+        Map<Integer, BigDecimal> calculate = EloCalculate.calculate(ContestType.RCR, context);
 
 
         logger.info(calculate.toString());
 
-        Map<Long, BigDecimal> ret = Map.of(
-                1L, new BigDecimal("1030.398"),
-                2L, new BigDecimal("1007.742"),
-                3L, new BigDecimal("989.91"),
-                4L, new BigDecimal("972.064")
+        Map<Integer, BigDecimal> ret = Map.of(
+                1, new BigDecimal("1030.398"),
+                2, new BigDecimal("1007.742"),
+                3, new BigDecimal("989.91"),
+                4, new BigDecimal("972.064")
         );
 
         Assert.equals(calculate, ret);
